Refactor the module
