+++ spec/lib/deprecations/version_spec.rb
