+++ spec/lib/deprecations_spec.rb
